requirements-client.txt
websocket-client==0.48.0

